pyproject.toml
[tool.poetry]
name = "diplom_django"
version = "0.1.0"
description = ""
authors = ["Yaroslav Galiulin <yaroslav66@list.ru>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.2.7"
djangorestframework = "^3.14.0"
python-dotenv = "^1.0.0"
djangorestframework-simplejwt = "^5.3.0"
psycopg2-binary = "^2.9.9"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
